requirements.txt
turtle
numpy
scikit-learn
pillow
